refacto(regulations): move AlertGroup in a new file
